refactored toolbar forms for re-usage #33
