+++ src/TimesheetBundle/Form/Toolbar/AbstractToolbarForm.php
@@ -1,0 +1,33 @@
+<?php
+
+/*
+ * This file is part of the Kimai package.
+ *
+ * (c) Kevin Papst <kevin@kevinpapst.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace TimesheetBundle\Form\Toolbar;
+
+use Symfony\Component\Form\AbstractType;
+
+/**
+ * Defines the base form used for all toolbars.
+ *
+ * @author Kevin Papst <kevin@kevinpapst.de>
+ */
+abstract class AbstractToolbarForm extends AbstractType
+{
+    /**
+     * Dirty hack to enable easy handling of GET form in controller and javascript.
+     *Cleans up the name of all form elents (and unfortunately of the form itself).
+     *
+     * @return null|string
+     */
+    public function getBlockPrefix()
+    {
+        return '';
+    }
+}
